% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Celsius to Kelvin conversion}
\usage{
celsius_to_kelvin(temp)
}
\arguments{
\item{temp}{A temperature expressed in Celsius}
}
\value{
A temperature in Kelvin
}
\description{
This function takes a temperature expressed in Celsius and convert it to Kelvin.
}
\examples{
kelvin_to_celsius(300)

}
