% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fahr_to_kelvin}
\alias{fahr_to_kelvin}
\title{Fahrenheit to Kelvin conversion}
\usage{
fahr_to_kelvin(temp)
}
\arguments{
\item{temp}{A temperature expressed in Fahrenheit}
}
\value{
A temperature in Kelvin
}
\description{
This function takes a temperature expressed in Fahrenheit and convert it to Kelvin.
}
\examples{
fahr_to_kelvin(32)

}
