% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot temperature trends}
\usage{
plot_trend(dat)
}
\arguments{
\item{dat}{A temperature dataframe}
}
\value{
A graphic
}
\description{
Plot temperature trends by month and estimate the average change in temperature across years.
}
\examples{
data("tuggeranong")
plot_trend(tuggeranong)

}
