% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Wide_To_Long_Temperature}
\alias{Wide_To_Long_Temperature}
\title{Wide to long conversion for standard monthly temperature}
\usage{
Wide_To_Long_Temperature(
  dat,
  id.vars = "Year",
  variable.name = "Month",
  value.name = "Temperature"
)
}
\arguments{
\item{dat}{A temperature dataframe}

\item{id.vars}{Character string. vector of id variables. Can be integer (variable position) or string (variable name). If blank, will use all non-measured variables.}

\item{variable.name}{Character string. Nname of variable used to store measured variable names}

\item{value.name}{Character string. Name of variable used to store values}
}
\value{
A dataframe
}
\description{
This function takes a standard monthly temperature dataframe in wide format to long format.
}
\examples{
data("tuggeranong")
Wide_To_Long_Temperature(tuggeranong)

}
